refactor: simplify point wrapped logic

+++ src/cartesian/point/point_u16/wrapped.rs
@@ -3,13 +3,13 @@ use crate::cartesian::point::point_i16::PointI16;
 use super::PointU16;
 
 pub fn assign_add(p: &mut PointU16, delta: &PointI16) {
-    p.x = p.x.wrapping_add(delta.x as i16 as u16);
-    p.y = p.y.wrapping_add(delta.y as i16 as u16);
+    p.x = p.x.wrapping_add(delta.x as u16);
+    p.y = p.y.wrapping_add(delta.y as u16);
 }
 
 pub fn add(p: &PointU16, delta: &PointI16) -> PointU16 {
-    let x = p.x.wrapping_add(delta.x as i16 as u16);
-    let y = p.y.wrapping_add(delta.y as i16 as u16);
+    let x = p.x.wrapping_add(delta.x as u16);
+    let y = p.y.wrapping_add(delta.y as u16);
     PointU16 { x, y }
 }
