+++ src/matrix/point/point_u8/wrapped.rs
@@ -3,13 +3,13 @@ use crate::matrix::point::point_i8::PointI8;
 use super::PointU8;
 
 pub fn assign_add(p: &mut PointU8, delta: &PointI8) {
-    p.row = p.row.wrapping_add(delta.row as i8 as u8);
-    p.col = p.col.wrapping_add(delta.col as i8 as u8);
+    p.row = p.row.wrapping_add(delta.row as u8);
+    p.col = p.col.wrapping_add(delta.col as u8);
 }
 
 pub fn add(p: &PointU8, delta: &PointI8) -> PointU8 {
-    let row = p.row.wrapping_add(delta.row as i8 as u8);
-    let col = p.col.wrapping_add(delta.col as i8 as u8);
+    let row = p.row.wrapping_add(delta.row as u8);
+    let col = p.col.wrapping_add(delta.col as u8);
     PointU8 { row, col }
 }
